pyproject.toml
[project]
name = "investmentframework"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pandas-datareader>=0.10.0",
]
